nearly all done

+++ edit/deleExp.php
@@ -1,8 +1,8 @@
 <?php
-// if(!isset($_SESSION["access"])){
-//     header("location /login.php");
-//     die("Login first!!!");
-// }
+if (!isset($_SESSION["access"])) {
+    header("location /login.php");
+    die("Login first!!!");
+}
 $id = $_REQUEST["q"];
 include('../db.php');
 $sql = "DELETE FROM expirences WHERE id = $id";
